removing useless test
